+++ core/app-version.js
@@ -1,1 +1,1 @@
-export const APP_VERSION = '20260315-1';
+export const APP_VERSION = '20260629-1';